chore: rework create group component, remove unused selectors
