--- src/js/app/selectors.js
@@ -1,6 +1,0 @@
-import { getLocation } from "connected-react-router";
-import { get } from "lodash-es";
-
-export const getRouterLocationState = state => getLocation(state).state;
-
-export const getRouterLocationStateValue = (state, key) => get(getRouterLocationState(state), key);
